+++ Terceiro_Modulo/LinguagemProgramacao_III/ExerciciosAvaliativos/exercicio-avaliativo-02/src/br/org/fundatec/model/Funcionario.java
@@ -4,12 +4,12 @@ import java.math.BigDecimal;
 
 public class Funcionario extends Pessoa {
 
-    private BigDecimal idFuncionario;
+    private Long idFuncionario;
     private BigDecimal salario;
     TipoFuncionario tipoFuncionario;
     Imobiliaria imobiliaria;
 
-    public Funcionario(BigDecimal idPessoa,String nomePessoa, Endereco endereco, String cpfPessoa, BigDecimal idFuncionario, BigDecimal salario, TipoFuncionario tipoFuncionario, Imobiliaria imobiliaria) {
+    public Funcionario(Long idPessoa,String nomePessoa, Endereco endereco, String cpfPessoa, Long idFuncionario, BigDecimal salario, TipoFuncionario tipoFuncionario, Imobiliaria imobiliaria) {
         super(idPessoa, nomePessoa, cpfPessoa, endereco);
         this.idFuncionario = idFuncionario;
         this.salario = salario;
@@ -17,11 +17,11 @@ public class Funcionario extends Pessoa {
         this.imobiliaria = imobiliaria;
     }
 
-    public BigDecimal getIdFuncionario() {
+    public Long getIdFuncionario() {
         return idFuncionario;
     }
 
-    public void setIdFuncionario(BigDecimal idFuncionario) {
+    public void setIdFuncionario(Long idFuncionario) {
         this.idFuncionario = idFuncionario;
     }
